requirements.txt
click==7.1.2
RPi.GPIO==0.7.0
